requirements.txt
django
gunicorn
djangorestframework
pillow
python-dotenv